svn: SvnChangeData renamed to ChangeData and moved to model core
SvnChangeSource was very generic and not specific to the Subversion
change source. This move allows the class to be used by other change
sources in the future.
