+++ de.setsoftware.reviewtool.core/src/de/setsoftware/reviewtool/model/changestructure/ChangestructureFactory.java
@@ -1,10 +1,15 @@
 package de.setsoftware.reviewtool.model.changestructure;
 
+import java.io.File;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 import de.setsoftware.reviewtool.model.api.IBinaryChange;
 import de.setsoftware.reviewtool.model.api.IChange;
+import de.setsoftware.reviewtool.model.api.IChangeData;
+import de.setsoftware.reviewtool.model.api.IChangeSource;
+import de.setsoftware.reviewtool.model.api.ICommit;
 import de.setsoftware.reviewtool.model.api.IFragment;
 import de.setsoftware.reviewtool.model.api.IHunk;
 import de.setsoftware.reviewtool.model.api.ILocalRevision;
@@ -66,5 +71,18 @@ public class ChangestructureFactory {
 
     public static IPositionInText createPositionInText(int line, int column) {
         return new PositionInText(line, column);
+    }
+
+    public static IChangeData createChangeData(
+            final IChangeSource source,
+            final List<? extends ICommit> commits) {
+        return new ChangeData(source, commits);
+    }
+
+    public static IChangeData createChangeData(
+            final IChangeSource source,
+            final List<? extends ICommit> commits,
+            final Map<File, IRevisionedFile> localPathMap) {
+        return new ChangeData(source, commits, localPathMap);
     }
 }
